feat: JWT auth, kanban, calendar, comments, file-upload, member-management, DB spec PPT

+++ PMS-all/PMS-auth-api/src/main/java/com/example/pmsauth/auth/LoginResponse.java
@@ -1,4 +1,4 @@
 package com.example.pmsauth.auth;
 
-public record LoginResponse(Long userId, String name, String token, String message) {
+public record LoginResponse(Integer userId, String name, String token, String message) {
 }
