+++ PMS-all/PMS-auth-api/src/main/java/com/example/pmsauth/auth/SignupResponse.java
@@ -1,4 +1,4 @@
 package com.example.pmsauth.auth;
 
-public record SignupResponse(Long userId, String name, String loginId) {
+public record SignupResponse(Integer userId, String name, String loginId) {
 }
